miniupnpd: Updated to 1.8-20140401

+++ release/src/router/miniupnpd/testasyncsendto.c
@@ -72,7 +72,7 @@ int test(void)
 		struct timeval timeout;
 		struct timeval now;
 		syslog(LOG_DEBUG, "get_next_scheduled_send : %d next_send=%ld.%06ld",
-		       i, next_send.tv_sec, next_send.tv_usec);
+		       i, (long)next_send.tv_sec, (long)next_send.tv_usec);
 		FD_ZERO(&writefds);
 		max_fd = 0;
 		gettimeofday(&now, NULL);
@@ -98,7 +98,7 @@ int test(void)
 		}
 		syslog(LOG_DEBUG, "get_sendto_fds() returned %d", i);
 		syslog(LOG_DEBUG, "select(%d, NULL, xx, NULL, %ld.%06ld)",
-		       max_fd, timeout.tv_sec, timeout.tv_usec);
+		       max_fd, (long)timeout.tv_sec, (long)timeout.tv_usec);
 		i = select(max_fd, NULL, &writefds, NULL, &timeout);
 		if(i < 0) {
 			syslog(LOG_ERR, "select: %m");
